devops-assessment/task-2/requirements.txt
Flask==2.1.2
azure-identity==1.7.0
azure-keyvault-secrets==4.3.0
